Partially made add company feature
