+++ backend/src/functions/jwt.ts
@@ -9,24 +9,3 @@ export const generateToken = async (payload: any) => {
 		expiresIn: "1d",
 	});
 };
-
-export const checkAuthenticated = ash(
-	async (req: Request, _, next: NextFunction) => {
-		const token = req.cookies.access_token;
-
-		if (!token) {
-			throw new MyError("Unauthorized", 401);
-		}
-
-		const decoded = jwt.verify(token, process.env.JWT_SECRET as string);
-
-		const { role } = decoded as any;
-
-		if (!decoded || role !== "admin") {
-			throw new MyError("Unauthorized", 401);
-		} else {
-			req.body.user = decoded;
-			next();
-		}
-	}
-);
