+++ webapp/src/pages/dashboard.tsx
@@ -1,0 +1,12 @@
+import { Navbar } from "@/components/Navbar";
+
+export default function Home() {
+	return (
+		<div className="w-screen h-screen bg-gray-100">
+			<Navbar />
+			<div className="w-full">
+
+			</div>
+		</div>
+	);
+}
